requirements.txt
arrow==1.3.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
binaryornot==0.4.4
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cookiecutter==2.6.0
crispy-tailwind==1.0.3
cycler==0.12.1
Django==4.2.13
django-browser-reload==1.12.1
django-cors-headers==4.4.0
django-crispy-forms==2.1
django-filter==24.2
django-tailwind==3.6.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
fontawesome-free==5.15.4
fontawesomefree==6.5.1
fonttools==4.53.0
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
networkx==3.1
numpy==1.24.4
opencv-python==4.10.0.82
packaging==24.1
pandas==2.0.3
pillow==10.3.0
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2023.7.10
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
zipp==3.19.2
